MA-22568 dma-buf/heaps: Correct the types of fd_flags and heap_flags
Casting both fd_flags and heap_flags into unsigned long.
The following problems will be encountered when the
system boot:

CFI failure at dma_heap_ioctl+0x210/0x3d8 (target: cma_uncached_heap_allocate+0x0/0x24 [cma_heap]; expected type: 0xec67cdc9)

+++ drivers/dma-buf/heaps/cma_heap.c
@@ -403,16 +403,16 @@ free_buffer:
 
 static struct dma_buf *cma_heap_allocate(struct dma_heap *heap,
 				  unsigned long len,
-				  unsigned long fd_flags,
-				  unsigned long heap_flags)
+				  u32 fd_flags,
+				  u64 heap_flags)
 {
 	return cma_heap_do_allocate(heap, len, fd_flags, heap_flags, false);
 }
 
 static struct dma_buf *cma_uncached_heap_allocate(struct dma_heap *heap,
 				  unsigned long len,
-				  unsigned long fd_flags,
-				  unsigned long heap_flags)
+				  u32 fd_flags,
+				  u64 heap_flags)
 {
 	return cma_heap_do_allocate(heap, len, fd_flags, heap_flags, true);
 }
@@ -420,8 +420,8 @@ static struct dma_buf *cma_uncached_heap_allocate(struct dma_heap *heap,
 /* Dummy function to be used until we can call coerce_mask_and_coherent */
 static struct dma_buf *cma_uncached_heap_not_initialized(struct dma_heap *heap,
 						unsigned long len,
-						unsigned long fd_flags,
-						unsigned long heap_flags)
+						u32 fd_flags,
+						u64 heap_flags)
 {
 	return ERR_PTR(-EBUSY);
 }
